refactor: separate logger interface and infrastructure implementation
Move logger interface to logger/logger.go and place concrete implementation in infra/log.go. This clarifies the dependency direction and separates abstraction from implementation. infra.go is still large and will be further split by responsibility.
